requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
attrs==24.3.0
bcrypt==4.2.1
black==24.8.0
boto3==1.35.25
botocore==1.35.25
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contextlib2==21.6.0
cryptography==43.0.1
dataclasses-json==0.6.7
diff-match-patch==20241021
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-flags==5.0.13
django-import-export==4.3.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-ninja==1.3.0
django-ninja-extra==0.21.8
django-ninja-jwt==5.3.4
django-storages==1.14.4
django-unfold==0.43.0
dnspython==2.6.1
duckduckgo_search==7.1.1
ecdsa==0.19.0
email_validator==2.1.1
Faker==30.1.0
fastapi==0.104.1
fastapi-cli==0.0.4
filelock==3.16.1
flake8==7.1.1
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.26.5
idna==3.8
injector==0.22.0
isort==5.13.2
Jinja2==3.1.4
jmespath==1.0.1
jsonpointer==2.1
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-text-splitters==0.0.2
langsmith==0.1.147
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.3
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
openai==1.39.0
orjson==3.10.4
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
primp==0.9.2
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-environ==0.4.54
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.17.32
s3transfer==0.10.2
safetensors==0.4.5
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.3
starlette==0.27.0
stripe==10.8.0
sympy==1.13.1
tablib==3.7.0
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.66.6
transformers==4.47.0
truststore==0.1.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.8.0
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.24.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wheel==0.41.2
yarl==1.18.3
zstandard==0.22.0
